chore(workflow): add release-please
this workflow will trigger when pushes (or merges) are made to main.

it will look through commits since the previous release and determine whether the next release is major, minor or patch.

it will update with the new version number and write the changelog.

the new release will be proposed as a PR.

upon merging the PR, the release will be created in Github.